pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-stacks"
version = "0.1.2"
description = "Installable skill packs that give AI agents deep domain expertise."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Agentic Stacks" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Systems Administration",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "click>=8.1",
    "fastapi>=0.115",
    "jinja2>=3.1",
    "slowapi>=0.1.9",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-tmp-files>=0.0.2",
    "pytest-asyncio>=0.24",
]
mcp = [
    "mcp[cli]>=1.0",
]
local = [
    "uvicorn[standard]>=0.32",
]

[project.urls]
Homepage = "https://agentic-stacks.com"
Repository = "https://github.com/agentic-stacks/agentic-stacks"
Documentation = "https://agentic-stacks.com/docs/getting-started"
Tracker = "https://github.com/agentic-stacks/agentic-stacks/issues"

[dependency-groups]
worker = [
    "workers-py",
    "workers-runtime-sdk",
]

[project.scripts]
agentic-stacks = "agentic_stacks_cli:main"
agentic-stacks-mcp = "agentic_stacks_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_stacks", "src/agentic_stacks_cli", "src/agentic_stacks_mcp", "src/web"]

[tool.pytest.ini_options]
testpaths = ["tests"]
